requirements.txt
click==5.1
pyusb==1.0.0
tabulate==0.7.5
six==1.10.0
pyusb==1.0.0